+++ ast/fuzz_test.go
@@ -2,9 +2,6 @@
 // Use of this source code is governed by an Apache2
 // license that can be found in the LICENSE file.
 
-//go:build go1.18
-// +build go1.18
-
 // nolint
 package ast
